Fix scan decision logic
For GitLab CI rules, if no rules match a job is not added to the pipeline.

In the case of release scans, we must ensure that the default behaviour is to
generate a scan job so as to not inadvertently release unscanned images.

This change adds a when: always rule and adds optional dependencies on
image jobs.

Signed-off-by: Evan Lezar <elezar@nvidia.com>